+++ querybook/server/logic/board.py
@@ -192,3 +192,8 @@ def get_boards_from_board_item(
 
 def update_es_boards_by_id(board_id: int):
     sync_elasticsearch.apply_async(args=[ElasticsearchItem.boards.value, board_id])
+
+
+@with_session
+def get_all_public_boards(session=None):
+    return session.query(Board).filter(Board.public is True).all()
