+++ querybook/webapp/components/BoardExpandableSection/BoardExpandableSection.scss
@@ -1,4 +1,4 @@
-.NavigatorBoardView {
+.BoardExpandableSection {
     background-color: var(--bg-light);
     border-radius: var(--border-radius-sm);
     overflow: hidden;
